// union-talk-frontend/src/services/generated/message/models/conversationRespVO.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { ConversationGroupInfoRespVO } from './conversationGroupInfoRespVO'
import type { ConversationUserInfoRespVO } from './conversationUserInfoRespVO'
import type { MessageRespVO } from './messageRespVO'

/**
 * 会话响应参数
 */
export interface ConversationRespVO {
  /** 用户会话id */
  userConversationId?: string
  /** 会话id */
  conversationId?: string
  /** 会话类型 */
  type?: string
  /** 群聊id */
  groupId?: string
  /** 最后一条消息id */
  lastMsgId?: string
  /** 最后一条消息时间 */
  lastMsgAt?: string
  /** 未读数量 */
  unreadCount?: number
  /** 最后已读消息id */
  lastReadMsgId?: string
  /** 是否置顶 */
  isPinned?: boolean
  /** 是否免打扰 */
  isMuted?: boolean
  /** 用户会话状态 */
  status?: string
  /** 用户会话更新时间 */
  updatedAt?: string
  /** 最后一条消息 */
  lastMessage?: MessageRespVO
  /** 私聊目标用户信息 */
  targetUser?: ConversationUserInfoRespVO
  /** 群聊信息 */
  groupInfo?: ConversationGroupInfoRespVO
}