// union-talk-frontend/src/services/generated/message/models/messageRespVO.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { ConversationUserInfoRespVO } from './conversationUserInfoRespVO'
import type { MessageMentionRespVO } from './messageMentionRespVO'

/**
 * 消息响应参数
 */
export interface MessageRespVO {
  /** 消息id */
  id?: string
  /** 会话id */
  conversationId?: string
  /** 发送者id */
  senderId?: string
  /** 发送者用户信息 */
  senderUser?: ConversationUserInfoRespVO
  /** 消息类型 */
  type?: string
  /** 消息内容 */
  content?: string
  /** 引用消息id */
  quoteMsgId?: string
  /** 结构化提及列表 */
  mentionList?: MessageMentionRespVO[]
  /** 是否撤回 */
  recalled?: boolean
  /** 创建时间 */
  createdAt?: string
}