// union-talk-frontend/src/services/generated/user/models/friendRequestRespVO.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

/**
 * 好友申请响应参数
 */
export interface FriendRequestRespVO {
  /** 申请id */
  id?: string
  /** 申请用户id */
  fromUserId?: string
  /** 申请用户名 */
  fromUsername?: string
  /** 申请用户头像 */
  fromAvatarUrl?: string
  /** 申请信息 */
  applyMsg?: string
  /** 申请状态 */
  status?: string
  /** 创建时间 */
  createdAt?: string
}