// union-talk-frontend/src/services/generated/user/models/friendRespVO.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

/**
 * 好友信息响应参数
 */
export interface FriendRespVO {
  /** 用户id */
  userId?: string
  /** 用户code */
  code?: string
  /** 用户名 */
  username?: string
  /** 头像地址 */
  avatarUrl?: string
  /** 备注 */
  remark?: string
  /** 好友分组id */
  friendGroupId?: string
  /** 关系状态 */
  status?: string
}