// union-talk-frontend/src/services/generated/user/models/groupMemberRespVO.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

/**
 * 群成员信息响应参数
 */
export interface GroupMemberRespVO {
  /** 用户id */
  userId?: string
  /** 用户名 */
  username?: string
  /** 头像地址 */
  avatarUrl?: string
  /** 群昵称 */
  nickname?: string
  /** 角色 */
  role?: string
  /** 加入时间 */
  joinedAt?: string
}