// union-talk-frontend/src/services/generated/user/models/userInfoRespVO.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { UserInfoRespVOStatus } from './userInfoRespVOStatus'

/**
 * 用户信息响应参数
 */
export interface UserInfoRespVO {
  /** 用户id */
  uid?: string
  /** 用户唯一CODE */
  code?: string
  /** 用户名 */
  username?: string
  /** 邮箱 */
  email?: string
  /** 头像地址 */
  avatarUrl?: string
  /** 简介 */
  bio?: string
  /** 添加好友是否需要验证 */
  needFriendVerify?: boolean
  /** 用户状态 */
  status?: UserInfoRespVOStatus
  /** 最近登录时间 */
  lastLoginAt?: string
}